central_server/requirements.txt
Django==2.2.9
djangorestframework==3.11.0
psycopg2==2.8.4
mysqlclient==1.4.6
